+++ test/abstractmodel.jl
@@ -3,6 +3,7 @@ using Test, ModelConstructors, UnPack
 m = GenericModel()
 m <= parameter(:a, 1., (0., 1.), (0., 1.), Untransformed(), fixed = false)
 m <= parameter(:b, 1., (0., 1.), (0., 1.), Untransformed(), fixed = false, scaling = x -> x / 100.)
+
 @testset "UnPack works with AbstractModel" begin
     @unpack a, b = m
     @test a == 1.