+++ test/parameters.jl
@@ -1,4 +1,4 @@
-using Test, ModelConstructors, Distributions, Dates, Nullables
+using Test, ModelConstructors, Distributions, Dates, Nullables, Random
 
 @testset "Ensure transformations to the real line/model space are valid" begin
     for T in subtypes(Transform)
@@ -120,6 +120,7 @@ end
     @test α.description == "No description available."
     @test α.tex_label == ""
     @test isa(α.transform, ModelConstructors.SquareRoot)
+    @test get_untransformed_values(α) == α.value
 end
 
 # UnscaledParameter, fixed = true
@@ -127,6 +128,7 @@ end
 @testset "Test fixed UnscaledParameter" begin
     @test α_fixed.transform_parameterization == (0.1596,0.1596)
     @test isa(α_fixed.transform, ModelConstructors.Untransformed)
+    @test get_untransformed_values(α_fixed) == α_fixed.value
 end
 
 # UnscaledParameter, fixed = true, transform should be overwritten given fixed
@@ -151,6 +153,7 @@ end
     @test isa(β, ScaledParameter)
     @test isa(β.prior.value, Gamma)
     @test isa(β.transform, ModelConstructors.Exponential)
+    @test get_untransformed_values(β) == β.scaledvalue
 end
 
 # Invalid transform
@@ -204,6 +207,35 @@ end
     global vals = [2., 2.]
     update!(pvec, vals, BitArray([true, false, true]))
     @test map(x -> x.value, pvec) == [1., 1., 1.]
+end
+
+@testset "Broadcasting with Parameter types" begin
+    a = parameter(:a, 1., (0., 1.), (0., 1.), Untransformed(), fixed = false)
+    b = parameter(:b, 1., (0., 1.), (0., 1.), Untransformed(), fixed = false, scaling = x -> x / 100.)
+
+    rand_vals = rand(3)
+    p=SteadyStateParameterGrid(:sspg, rand(3))
+    p2=SteadyStateParameterGrid(:sspg, rand(3,2))
+    rand1 = rand()
+
+    @test a .* ones(3) == a.value * ones(3)
+    @test a .* rand_vals == a.value * rand_vals
+    @test b .* ones(3) == b.scaledvalue * ones(3)
+    @test b .* rand_vals == b.scaledvalue * rand_vals
+    @test p .* 1 == p.value
+    @test p .* rand1 == p.value .* rand1
+    @test p .* rand_vals == p.value .* rand_vals
+    @test a .* ones(3) == a.value * ones(3)
+    @test a .+ rand_vals == a.value .+ rand_vals
+    @test b .+ ones(3) == b.scaledvalue .+ ones(3)
+    @test b .+ rand_vals == b.scaledvalue .+ rand_vals
+    @test p .+ 1 == p.value .+ 1.
+    @test p .+ rand1 == p.value .+ rand1
+    @test p .+ rand_vals == p.value + rand_vals
+    @test p[1] == p.value[1]
+    @test first(p[1:1]) == p.value[1]
+    @test p[2:3] == p.value[2:3]
+    @test -p == -p.value
 end
 
 nothing